Remove AppTestCase and move pieces into individual files
